refactor: modify newcarracing package based on step5 requirement
