+++ src/main/java/newcarracing/domain/GameRule.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 public class GameRule {
     private RuleCondition ruleCondition;
