+++ src/main/java/newcarracing/domain/RacingCar.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 import java.util.Objects;
