+++ src/main/java/newcarracing/domain/RacingCars.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 import java.util.ArrayList;
 import java.util.List;
