+++ src/main/java/newcarracing/domain/RandomRuleCondition.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 import java.util.Random;
