+++ src/main/java/newcarracing/domain/RuleCondition.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 public interface RuleCondition {
     boolean checkCondition();
