+++ src/test/java/newcarracing/domain/RacingCarTest.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
