+++ src/test/java/newcarracing/domain/RacingCarsTest.java
@@ -1,4 +1,4 @@
-package newcarracing;
+package newcarracing.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
