+++ src/test/java/newcarracing/domain/RacingGameTest.java
@@ -1,5 +1,6 @@
-package newcarracing;
+package newcarracing.domain;
 
+import newcarracing.view.ResultView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
